maybe?

+++ Joint.java
@@ -1,8 +1,8 @@
 package edu.mccc.cos210.fp.pupp;
 
 public class Joint {
-	protected int x;
-	protected int y;
+	protected double x;
+	protected double y;
 	protected ILimb upperLimb;
 	protected ILimb lowerLimb;
 	
@@ -19,16 +19,16 @@ public class Joint {
 		this(x,y,upperLimb,null);
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
